chore(stream-tile-handler, stream-model-handler, stream-model-instance-handler): Unify implementation of signature verification across streamtypes
