+++ packages/common/src/index.ts
@@ -9,6 +9,7 @@ export * from './utils/test-utils.js'
 export * from './utils/accountid-utils.js'
 export * from './utils/cid-utils.js'
 export * from './utils/abort-signal-utils.js'
+export * from './utils/signature_utils.js'
 export * from './logger-provider.js'
 export * from './loggers.js'
 export * from './networks.js'
